src/requirements.txt
mysql-connector-python==8.0.26
openpyxl==3.0.7
psycopg2-binary==2.9.1

# VSCode
autopep8
pylint
rope
